+++ poteto.go
@@ -59,7 +59,9 @@ func (p *poteto) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// Get And SetRequestId
 	reqId := ctx.RequestId()
 	ctx.Set(constant.STORE_REQUEST_ID, reqId)
-	ctx.GetResponse().Header().Set(constant.HEADER_X_REQUEST_ID, reqId)
+	if id := ctx.GetRequest().Header.Get(constant.HEADER_X_REQUEST_ID); id == "" {
+		ctx.GetResponse().Header().Set(constant.HEADER_X_REQUEST_ID, reqId)
+	}
 
 	routes := p.router.GetRoutesByMethod(r.Method)